Add H3ErrorCode
This is a more-or-less direct translation between a C enum and a Swift
enum that attempts to preserve the information available in the source
code comments. It adds one new error code, SizeInvalid, to capture a few
preconditions checking sizes. This may not be the best way to handle it,
but it's better than a precondition for now.